Move CatalogPde into its own class so that it is easy to link to so that the DSL stays up-to-date.
