+++ plugin-gradle/src/main/java/dev/equo/ide/gradle/P2ModelDslWithCatalog.java
@@ -15,6 +15,7 @@ package dev.equo.ide.gradle;
 
 import dev.equo.ide.Catalog;
 import dev.equo.ide.CatalogDsl;
+import dev.equo.ide.CatalogPde;
 import dev.equo.ide.IdeHook;
 import dev.equo.ide.IdeHookBuildship;
 import java.io.File;
@@ -107,7 +108,7 @@ public class P2ModelDslWithCatalog extends P2ModelDsl {
 
 		/** Ignore / Error */
 		public void missingApiBaseline(String ignoreOrError) {
-			Catalog.Pde.missingApiBaseline(workspaceInit(), ignoreOrError);
+			CatalogPde.missingApiBaseline(workspaceInit(), ignoreOrError);
 		}
 	}
