+++ plugin-maven/src/main/java/dev/equo/ide/maven/AbstractP2MojoWithCatalog.java
@@ -15,6 +15,7 @@ package dev.equo.ide.maven;
 
 import dev.equo.ide.Catalog;
 import dev.equo.ide.CatalogDsl;
+import dev.equo.ide.CatalogPde;
 import dev.equo.ide.IdeHook;
 import dev.equo.ide.IdeHookM2E;
 import dev.equo.ide.WorkspaceInit;
@@ -66,7 +67,7 @@ public abstract class AbstractP2MojoWithCatalog extends AbstractP2Mojo {
 		protected void processVersionOverrides() {
 			super.processVersionOverrides();
 			if (missingApiBaseline != null) {
-				Catalog.Pde.missingApiBaseline(workspaceInit(), missingApiBaseline);
+				CatalogPde.missingApiBaseline(workspaceInit(), missingApiBaseline);
 			}
 		}
 	}
